Improvements to hotplugging on Desktop.
Changes JamepadControllerMonitor to track controllers by SDL instance ID instead of SDL index. This keeps the association of Controller java objects and physical devices constant.

Cuts out the check for new controllers in JamepadControllerMonitor by instead updating the controller list when Jamepad reports that its controller list was updated.

Adds the ability for JamepadController to report its name when disconnected by storing the name on connect.

Allows JampadController to treat a null ControllerIndex as a disconnected controller.

Fixes a bug in JamepadControllerManager that caused the controller update code to run twice as often as intended.

Fixes a bug in ControllersTest that caused the first controller to be unresponsive when the program first starts.

Fixes a bug in ControllersTest that caused duplicate disconnect messages in sone circumstances.

Updates to the newest snapshot of Jamepad.